refactor: add error handling in queries
