+++ src/api/query/auth-query.ts
@@ -1,8 +1,7 @@
 'use server';
 
-import { API_HOST } from '@/api';
+import { API_HOST, ApiResponse, createApiFetchError, interceptResponse } from '@/api';
 import { getCommonFetchConfig } from '@/api/config';
-import { ApiResponse } from '@/api/type/common.type';
 
 type Vendor = 'GOOGLE' | 'NAVER' | 'APPLE';
 
@@ -12,19 +11,15 @@ export interface GetMyInfoQueryReturn {
   vendor: Vendor;
   createdAt: Date;
 }
+
 export const getMyInfoQuery = async (): Promise<ApiResponse<GetMyInfoQueryReturn>> => {
   try {
-    const data = await fetch(`${API_HOST}/auth/me`, {
+    const response = await fetch(`${API_HOST}/auth/me`, {
       method: 'GET',
       ...(await getCommonFetchConfig())
     });
-    return data.json();
+    return interceptResponse(response);
   } catch (error) {
-    return {
-      result: undefined,
-      error: {
-        code: 'FETCH_ERROR'
-      }
-    };
+    return createApiFetchError();
   }
 };
