+++ src/api/query/diary-query.ts
@@ -1,10 +1,9 @@
 'use server';
 
-import { API_HOST } from '@/api';
+import { API_HOST, ApiResponse, createApiFetchError, interceptResponse } from '@/api';
 import { getCommonFetchConfig } from '@/api/config';
-import { ApiResponse } from '@/api/type/common.type';
 
-export type GetAllDiaryQueryReturn = {
+export interface GetAllDiaryQueryReturn {
   readonly content: Array<{
     readonly diaryId: number;
     readonly title: string;
@@ -17,13 +16,18 @@ export type GetAllDiaryQueryReturn = {
     readonly totalElements: number;
     readonly totalPages: number;
   };
-};
+}
+
 export const getAllDiaryQuery = async (size: number): Promise<ApiResponse<GetAllDiaryQueryReturn>> => {
-  const data = await fetch(`${API_HOST}/diary?page=0&size=${size}`, {
-    method: 'GET',
-    ...(await getCommonFetchConfig())
-  });
-  return data.json();
+  try {
+    const response = await fetch(`${API_HOST}/diary?page=0&size=${size}`, {
+      method: 'GET',
+      ...(await getCommonFetchConfig())
+    });
+    return interceptResponse(response);
+  } catch (error) {
+    return createApiFetchError();
+  }
 };
 
 export interface GetDiaryQueryReturn {
@@ -33,10 +37,15 @@ export interface GetDiaryQueryReturn {
   readonly createdAt: Date;
   readonly updatedAt: Date;
 }
+
 export const getDiaryQuery = async (id: number): Promise<ApiResponse<GetDiaryQueryReturn>> => {
-  const data = await fetch(`${API_HOST}/diary/${id}`, {
-    method: 'GET',
-    ...(await getCommonFetchConfig())
-  });
-  return data.json();
+  try {
+    const data = await fetch(`${API_HOST}/diary/${id}`, {
+      method: 'GET',
+      ...(await getCommonFetchConfig())
+    });
+    return interceptResponse(data);
+  } catch (error) {
+    return createApiFetchError();
+  }
 };
